New PECL-free fsevents watcher implementation.
In the transition to PHP 7, the API for writing C-based PHP extension
(PECL) changed quite a bit to accommodate the change in the architecture
of the PHP engine.  This broke the fsevents PECL extension.  I was able
to get it building again but was getting difficult to diagnose segfaults
during execution.

As a workaround, I wrote a small C program that just uses the Mac OS
CoreServices framework to handle the fsevents monitoring.  When an event
is detected, it just outputs the monitored path that was affected by the
filesystem change.

On the PHP side, I use React PHP to watch the output of the C process.
When I see a path, I call the associated callback.  This gets us the
filesystem monitoring capabilities without needing an actual PECL
extension.  Because the OS-level APIs are very stable, we can likely
handle filesystem monitoring without requiring any kind of compilation
step going forward, which will also improve the user experience for Mac
folks.